Simulate core logic
Technically, I could remove the echo in line 90 and wrap it up. However,
I think this tool should be a little bit more opinionated and it should
be used.

These opinions could be either on its logic or documentation, I'm still
pondering what could be the best. But that's a problem for "tomorrow's
me".

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>