Works as expected
On the last commit, I wasn't 100% sure how to call the script and what
the minimal "opinionated" behavior was.

As it is right now, it will require from the user a local git
repository, the gh and git binaries, and the GH_TOKEN env. variable.

I believe "it just works", but I don't feel it is "bullet proof" yet.

PS: I was able to generate shadows from my deactivated
jossemargt-3pillar GH account :D

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>